+++ packages/visual-stack/src/components/tests/CollapsiblePanel.test.js
@@ -1,104 +1,90 @@
 import React from 'react';
-import { CollapsiblePanel } from '../CollapsiblePanel';
-import { shallow } from 'enzyme';
+import { CollapsiblePanel, PureCollapsiblePanel } from '../CollapsiblePanel';
+import { mount } from 'enzyme';
 
 import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 Enzyme.configure({ adapter: new Adapter() });
 
 describe('CollapsiblePanel', () => {
-  let shallowWrapper;
-  let onToggleCollapsed;
+  let container;
 
   describe('when initially collapsed', () => {
     beforeEach(() => {
-      onToggleCollapsed = jest.fn();
-      shallowWrapper = shallow(
-        <CollapsiblePanel onToggleCollapsed={onToggleCollapsed}>
+      container = mount(
+        <CollapsiblePanel>
           <div id="test-marker">I am some content</div>
         </CollapsiblePanel>
       );
     });
 
     test('should not initially show content', () => {
-      expect(shallowWrapper.find('div#test-marker').length).toBe(0);
+      expect(container.find('div#test-marker').length).toBe(0);
     });
 
     test('should toggle content when button is clicked', () => {
-      const headerButton = shallowWrapper.find(
-        '.vs-collapsible-panel-header-button'
+      const headerButton = container.find(
+        'button.vs-collapsible-panel-header-button'
       );
-      expect(shallowWrapper.find('div#test-marker').length).toBe(0);
+      expect(container.find('div#test-marker').length).toBe(0);
       headerButton.simulate('click');
-      expect(shallowWrapper.find('div#test-marker').length).toBe(1);
-      expect(onToggleCollapsed.mock.calls[0][0]).toEqual(false);
+      expect(container.find('div#test-marker').length).toBe(1);
       headerButton.simulate('click');
-      expect(shallowWrapper.find('div#test-marker').length).toBe(0);
-      expect(onToggleCollapsed.mock.calls[1][0]).toEqual(true);
+      expect(container.find('div#test-marker').length).toBe(0);
     });
-    
+
     test('should toggle content when title is clicked', () => {
-      const headerTitle = shallowWrapper.find(
-        '.vs-collapsible-panel-header-title'
+      const headerTitle = container.find(
+        'button.vs-collapsible-panel-header-button'
       );
-      expect(shallowWrapper.find('div#test-marker').length).toBe(0);
+      expect(container.find('div#test-marker').length).toBe(0);
       headerTitle.simulate('click');
-      expect(shallowWrapper.find('div#test-marker').length).toBe(1);
-      expect(onToggleCollapsed.mock.calls[0][0]).toEqual(false);
+      expect(container.find('div#test-marker').length).toBe(1);
       headerTitle.simulate('click');
-      expect(shallowWrapper.find('div#test-marker').length).toBe(0);
-      expect(onToggleCollapsed.mock.calls[1][0]).toEqual(true);
+      expect(container.find('div#test-marker').length).toBe(0);
     });
   });
 
   describe('when not initially collapsed', () => {
     beforeEach(() => {
-      onToggleCollapsed = jest.fn();
-      shallowWrapper = shallow(
-        <CollapsiblePanel
-          onToggleCollapsed={onToggleCollapsed}
-          initializedCollapsed={false}
-        >
+      container = mount(
+        <CollapsiblePanel initializedCollapsed={false} >
           <div id="test-marker">I am some content</div>
         </CollapsiblePanel>
       );
     });
 
     test('should initially show the content', () => {
-      expect(shallowWrapper.find('div#test-marker').length).toBe(1);
+      expect(container.find('div#test-marker').length).toBe(1);
     });
 
     test('should toggle content when button is clicked', () => {
-      const headerButton = shallowWrapper.find(
-        '.vs-collapsible-panel-header-button'
+      const headerButton = container.find(
+        'button.vs-collapsible-panel-header-button'
       );
-      expect(shallowWrapper.find('div#test-marker').length).toBe(1);
+      expect(container.find('div#test-marker').length).toBe(1);
       headerButton.simulate('click');
-      expect(shallowWrapper.find('div#test-marker').length).toBe(0);
-      expect(onToggleCollapsed.mock.calls[0][0]).toEqual(true);
+      expect(container.find('div#test-marker').length).toBe(0);
       headerButton.simulate('click');
-      expect(shallowWrapper.find('div#test-marker').length).toBe(1);
-      expect(onToggleCollapsed.mock.calls[1][0]).toEqual(false);
+      expect(container.find('div#test-marker').length).toBe(1);
     });
 
     test('should toggle content when title is clicked', () => {
-      const headerTitle = shallowWrapper.find(
-        '.vs-collapsible-panel-header-title'
+      const headerTitle = container.find(
+        'button.vs-collapsible-panel-header-button'
       );
-      expect(shallowWrapper.find('div#test-marker').length).toBe(1);
+      expect(container.find('div#test-marker').length).toBe(1);
       headerTitle.simulate('click');
-      expect(shallowWrapper.find('div#test-marker').length).toBe(0);
-      expect(onToggleCollapsed.mock.calls[0][0]).toEqual(true);
+      expect(container.find('div#test-marker').length).toBe(0);
       headerTitle.simulate('click');
-      expect(shallowWrapper.find('div#test-marker').length).toBe(1);
-      expect(onToggleCollapsed.mock.calls[1][0]).toEqual(false);
+      expect(container.find('div#test-marker').length).toBe(1);
     });
   });
 
   describe('when rendering icon', () => {
     it('should render icon when I pass in an icon prop', () => {
       const titleIcon = <div id="titleIcon" />;
-      const container = shallow(
+      const container = mount(
         <CollapsiblePanel titleIcon={titleIcon} initializedCollapsed={false} />
       );
       expect(
@@ -110,8 +96,116 @@ describe('CollapsiblePanel', () => {
     });
 
     it('should not render icon when I dont pass in an icon prop', () => {
-      const container = shallow(
+      const container = mount(
         <CollapsiblePanel initializedCollapsed={false} />
+      );
+      expect(
+        container.find('.vs-collapsible-panel-icon-container #titleIcon')
+      ).toHaveLength(0);
+      expect(
+        container.find('.vs-collapsible-panel-icon-placeholder')
+      ).toHaveLength(0);
+    });
+  });
+});
+
+describe('PureCollapsiblePanel', () => {
+  let container;
+  let onChange;
+
+  describe('when collapsed prop is true', () => {
+
+    test('render collapsed and respond false when button clicked', () => {
+      onChange = jest.fn();
+      container = mount(
+        <PureCollapsiblePanel collapsed={true} onChange={onChange}>
+          <div id="test-marker">I am some content</div>
+        </PureCollapsiblePanel>
+      );
+
+      const headerButton = container.find(
+        'button.vs-collapsible-panel-header-button'
+      );
+      expect(container.find('div#test-marker').length).toBe(0);
+      headerButton.simulate('click');
+      expect(container.find('div#test-marker').length).toBe(0);
+      expect(onChange.mock.calls[0][0]).toEqual(false);
+    });
+
+    test('render collapsed and respond false when title clicked', () => {
+      onChange = jest.fn();
+      container = mount(
+        <PureCollapsiblePanel collapsed={true} onChange={onChange}>
+          <div id="test-marker">I am some content</div>
+        </PureCollapsiblePanel>
+      );
+
+      const headerButton = container.find(
+        '.vs-collapsible-panel-header-title'
+      );
+      expect(container.find('div#test-marker').length).toBe(0);
+      headerButton.simulate('click');
+      expect(container.find('div#test-marker').length).toBe(0);
+      expect(onChange.mock.calls[0][0]).toEqual(false);
+    });
+
+
+  });
+
+  describe('when collapsed prop is false', () => {
+
+    test('render expanded and respond true when button clicked', () => {
+      onChange = jest.fn();
+      container = mount(
+        <PureCollapsiblePanel collapsed={false} onChange={onChange}>
+          <div id="test-marker">I am some content</div>
+        </PureCollapsiblePanel>
+      );
+
+      const headerButton = container.find(
+        'button.vs-collapsible-panel-header-button'
+      );
+      expect(container.find('div#test-marker').length).toBe(1);
+      headerButton.simulate('click');
+      expect(container.find('div#test-marker').length).toBe(1);
+      expect(onChange.mock.calls[0][0]).toEqual(true);
+    });
+
+    test('render expanded and respond true when title clicked', () => {
+      onChange = jest.fn();
+      container = mount(
+        <PureCollapsiblePanel collapsed={false} onChange={onChange}>
+          <div id="test-marker">I am some content</div>
+        </PureCollapsiblePanel>
+      );
+
+      const headerButton = container.find(
+        '.vs-collapsible-panel-header-title'
+      );
+      expect(container.find('div#test-marker').length).toBe(1);
+      headerButton.simulate('click');
+      expect(container.find('div#test-marker').length).toBe(1);
+      expect(onChange.mock.calls[0][0]).toEqual(true);
+    });
+  });
+
+  describe('when rendering icon', () => {
+    it('should render icon when I pass in an icon prop', () => {
+      const titleIcon = <div id="titleIcon" />;
+      const container = mount(
+        <PureCollapsiblePanel titleIcon={titleIcon} collapsed={false} onChange={()=>{}} />
+      );
+      expect(
+        container.find('.vs-collapsible-panel-icon-container #titleIcon')
+      ).toHaveLength(1);
+      expect(
+        container.find('.vs-collapsible-panel-icon-placeholder')
+      ).toHaveLength(1);
+    });
+
+    it('should not render icon when I dont pass in an icon prop', () => {
+      const container = mount(
+        <PureCollapsiblePanel collapsed={false} onChange={()=>{}}/>
       );
       expect(
         container.find('.vs-collapsible-panel-icon-container #titleIcon')